tests/unit_tests/pdf_processing_/ocr_engine_/test_ocr_engine_abstract_base.py
# Test file for TestOCREngineAbstractBase
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
# Test suite for ipfs_datasets_py.pdf_processing.ocr_engine

import os
import io
import time
import threading
from abc import ABC
from unittest.mock import Mock, patch, MagicMock

import pytest
import numpy as np
from PIL import Image


import cv2
import torch
import pytesseract

from ipfs_datasets_py.pdf_processing.ocr_engine import (
    OCREngine,
    EasyOCR,
    SuryaOCR,
    TesseractOCR,
    TrOCREngine,
    MultiEngineOCR
)

from .concrete_ocr_engine import ConcreteOCREngine

from ipfs_datasets_py.pdf_processing.ocr_engine import OCREngine
from tests._test_utils import (
    has_good_callable_metadata,
    raise_on_bad_callable_code_quality,
    get_ast_tree,
    BadDocumentationError,
    BadSignatureError
)

work_dir = "/home/runner/work/ipfs_datasets_py/ipfs_datasets_py"
file_path = os.path.join(work_dir, "ipfs_datasets_py/pdf_processing/ocr_engine.py")
md_path = os.path.join(work_dir, "ipfs_datasets_py/pdf_processing/ocr_engine_stubs.md")

# Make sure the input file and documentation file exist.
assert os.path.exists(file_path), f"Input file does not exist: {file_path}. Check to see if the file exists or has been moved or renamed."
assert os.path.exists(md_path), f"Documentation file does not exist: {md_path}. Check to see if the file exists or has been moved or renamed."

from ipfs_datasets_py.pdf_processing.ocr_engine import (
    EasyOCR,
    MultiEngineOCR,
    OCREngine,
    SuryaOCR,
    TesseractOCR,
    TrOCREngine
)

# Check if each classes methods are accessible:
assert OCREngine._initialize
assert OCREngine.extract_text
assert OCREngine.is_available
assert SuryaOCR._initialize
assert SuryaOCR.extract_text
assert TesseractOCR._initialize
assert TesseractOCR.extract_text
assert TesseractOCR._preprocess_image
assert EasyOCR._initialize
assert EasyOCR.extract_text
assert TrOCREngine._initialize
assert TrOCREngine.extract_text
assert MultiEngineOCR.extract_with_ocr
assert MultiEngineOCR.get_available_engines
assert MultiEngineOCR.classify_document_type


# Check if the file's imports are accessible
try:
    from abc import ABC, abstractmethod
    import logging
    import io
    import numpy as np
    from typing import Dict, List, Any
    from PIL import Image
    import cv2
except ImportError as e:
    raise ImportError(f"Failed to import necessary modules: {e}")





class TestOCREngineAbstractBase:
    """Test suite for OCREngine abstract base class."""

    def test_ocr_engine_is_abstract_class(self):
        """
        GIVEN the OCREngine class
        WHEN attempting to instantiate it directly
        THEN should raise TypeError due to abstract methods
        """
        with pytest.raises(TypeError, match="Can't instantiate abstract class"):
            OCREngine("test")

    def test_ocr_engine_inheritance_structure(self):
        """
        GIVEN the OCREngine class
        WHEN examining its inheritance
        THEN should inherit from ABC (Abstract Base Class)
        AND should define abstract methods _initialize and extract_text
        """
        # Should inherit from ABC
        assert issubclass(OCREngine, ABC)
        
        # Should have abstract methods
        abstract_methods = OCREngine.__abstractmethods__
        expected_abstract = {'_initialize', 'extract_text'}
        assert expected_abstract.issubset(abstract_methods)

    def test_ocr_engine_init_sets_name_and_available(self):
        """
        GIVEN a concrete OCREngine subclass with valid name
        WHEN initializing with __init__(name)
        THEN should set self.name to provided name
        AND should set self.available to False initially
        AND should call _initialize() method
        """
        engine = ConcreteOCREngine("test_engine")
        
        assert engine.name == "test_engine"
        assert engine.available == True  # Set to True by our test _initialize
        
    def test_ocr_engine_init_calls_initialize(self):
        """
        GIVEN a concrete OCREngine subclass
        WHEN initializing
        THEN should call _initialize() method during construction
        """
        with patch.object(ConcreteOCREngine, '_initialize') as mock_init:
            ConcreteOCREngine("test")
            mock_init.assert_called_once()

    def test_ocr_engine_init_with_empty_name(self):
        """
        GIVEN a concrete OCREngine subclass
        WHEN initializing with empty string name
        THEN should handle gracefully and set name to empty string
        """
        engine = ConcreteOCREngine("")
        assert engine.name == ""
        assert hasattr(engine, 'available')

    def test_ocr_engine_init_with_none_name(self):
        """
        GIVEN a concrete OCREngine subclass
        WHEN initializing with None as name
        THEN should raise TypeError
        """
        with pytest.raises(TypeError):
            ConcreteOCREngine(None)

    def test_ocr_engine_init_handles_initialize_failure(self):
        """
        GIVEN a concrete OCREngine subclass that fails initialization
        WHEN initializing
        THEN should handle _initialize() failures gracefully
        AND should set available to False
        """
        engine = ConcreteOCREngine("test", should_fail_init=True)
        # The engine should handle the failure and set available appropriately
        # This depends on implementation - it might stay False or be set to False
        assert hasattr(engine, 'available')
        assert hasattr(engine, 'name')
        assert engine.name == "test"

    def test_is_available_returns_boolean(self):
        """
        GIVEN an initialized OCREngine instance
        WHEN calling is_available()
        THEN should return boolean value
        AND should reflect the current available status
        """
        # Test with available engine
        engine = ConcreteOCREngine("test")
        result = engine.is_available()
        assert isinstance(result, bool)
        assert result == engine.available
        
        # Test with unavailable engine
        engine_fail = ConcreteOCREngine("test_fail", should_fail_init=True)
        result_fail = engine_fail.is_available()
        assert isinstance(result_fail, bool)

    def test_is_available_thread_safety(self):
        """
        GIVEN an OCREngine instance
        WHEN calling is_available() from multiple threads simultaneously
        THEN should be thread-safe and return consistent results
        """
        engine = ConcreteOCREngine("test")
        results = []
        exceptions = []
        
        def check_availability():
            try:
                for _ in range(100):
                    result = engine.is_available()
                    results.append(result)
                    time.sleep(0.001)  # Small delay to encourage race conditions
            except Exception as e:
                exceptions.append(e)
        
        threads = [threading.Thread(target=check_availability) for _ in range(5)]
        
        for thread in threads:
            thread.start()
        
        for thread in threads:
            thread.join()
        
        # Should have no exceptions
        assert len(exceptions) == 0, f"Thread safety test failed with exceptions: {exceptions}"
        
        # All results should be consistent (same boolean value)
        assert len(results) > 0
        first_result = results[0]
        assert all(r == first_result for r in results), "is_available() returned inconsistent results across threads"

    def test_abstract_methods_must_be_implemented(self):
        """
        GIVEN a class that inherits from OCREngine
        WHEN the class doesn't implement abstract methods
        THEN should not be instantiable
        """
        class IncompleteEngine(OCREngine):
            def _initialize(self):
                pass
            # Missing extract_text implementation
        
        with pytest.raises(TypeError, match="Can't instantiate abstract class"):
            IncompleteEngine("incomplete")
        
        class AnotherIncompleteEngine(OCREngine):
            def extract_text(self, image_data: bytes):
                return {}
            # Missing _initialize implementation
        
        with pytest.raises(TypeError, match="Can't instantiate abstract class"):
            AnotherIncompleteEngine("incomplete2")

    def test_concrete_implementation_works(self):
        """
        GIVEN a properly implemented concrete OCREngine subclass
        WHEN instantiating and using it
        THEN should work correctly
        """
        engine = ConcreteOCREngine("working_engine")
        
        # Should be available after successful initialization
        assert engine.is_available()
        
        # Should be able to extract text
        result = engine.extract_text(b"fake_image_data")
        assert isinstance(result, dict)
        assert 'text' in result
        assert 'confidence' in result
        assert 'engine' in result
        assert result['engine'] == 'working_engine'

    def test_engine_name_attribute_immutable_after_init(self):
        """
        GIVEN an initialized OCREngine instance
        WHEN accessing the name attribute
        THEN should return the name set during initialization
        AND should not change unexpectedly
        """
        engine = ConcreteOCREngine("test_name")
        original_name = engine.name
        
        # Name should be what we set
        assert engine.name == "test_name"
        
        # Try to modify (this tests that the name stays consistent)
        # We're not testing immutability enforcement, just consistency
        assert engine.name == original_name

    def test_available_attribute_reflects_initialization_state(self):
        """
        GIVEN OCREngine instances with different initialization outcomes
        WHEN checking the available attribute
        THEN should accurately reflect whether initialization succeeded
        """
        # Successful initialization
        engine_success = ConcreteOCREngine("success")
        assert engine_success.available == True
        
        # Failed initialization
        engine_fail = ConcreteOCREngine("fail", should_fail_init=True)
        # The available state should be False when initialization fails
        assert engine_fail.available == False
        assert isinstance(engine_fail.available, bool)



if __name__ == "__main__":
    pytest.main([__file__, "-v"])
