requirements.txt
# Core dependencies only - for initial testing
# Use this to test if basic setup works, then add more dependencies gradually

# Local submodule dependencies
# -e ./ipfs_kit_py  # Commented out - directory is empty and not properly configured

# Core Python packages
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
requests>=2.25.0,<3.0.0

# Caching for CLI tools
cachetools>=5.3.0,<6.0.0

# Testing framework
pytest>=7.3.1,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-parallel>=0.1.1 # Allows parallel test execution
pytest-xdist>=3.8.0
pytest-timeout >=2.0.2 # To prevent test suites from hanging indefinitely


# Basic utilities
tqdm>=4.60.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
nltk>=3.8.1

# Development tools
coverage>=7.0.0,<8.0.0
mypy>=1.0.0,<2.0.0
flake8>=6.0.0,<8.0.0

# Data processing
pyarrow>=15.0.0,<21.0.0
jsonschema>=4.0.0,<5.0.0
pydantic>=2.0.0,<3.0.0
jsonnet>=0.20.0  # Jsonnet data templating language support

# PDF processing
networkx>=3.1
pdfplumber>=0.11.7
pytesseract>=0.3.13
tiktoken>=0.6.0
opencv-contrib-python-headless>=4.11.0.86 # NOTE THis library errors with "Clock skew detected"

Flask>=3.1.1
duckdb>=1.3.2
pymupdf>=1.26.3
yt-dlp>=2025.6.30
surya-ocr>=0.14.6
pillow>=10.4.0
ffmpeg-python>=0.2.0
openai>=1.97.1
toml>=0.10.2

# Model Context Protocol (MCP)
# Provides `mcp.server.FastMCP` and `mcp.types` used by the MCP server
mcp>=1.2.0

# System monitoring utilities
psutil>=5.9.0

# For testing only
Faker>=37.4.2
reportlab>=4.4.3
pyfakefs
moviepy
cryptography # For encrypting/decrypting test files

# ipld crap
ipld-car
ipld-dag-pb
dag-cbor>=0.3.3

# All this other crap in order to get the debugger to fucking work.
slugify
llama-index>=0.13.5
qdrant-client
elasticsearch
pysbd
PyJWT
fastapi
keyring
passlib
uvicorn
psutil

# Web scraping and archiving tools
cdx-toolkit>=0.9.37  # Common Crawl CDX access
wayback>=0.4.5  # Wayback Machine API
internetarchive>=5.5.0  # Internet Archive access
autoscraper>=1.1.14  # Intelligent web scraping
ipwb>=0.2024.10.24  # InterPlanetary Wayback Machine
warcio>=1.7.4  # WARC file processing
beautifulsoup4>=4.12.0  # HTML parsing for scraping
selenium>=4.15.0  # Browser automation for dynamic content
scrapy>=2.11.0  # Advanced web scraping framework
aiohttp>=3.8.0  # Async HTTP client for web archiving

# Add these one by one after basic setup works:
datasets>=2.10.0,<3.0.0
# Vector store backend
faiss-cpu>=1.8.0
# transformers>=4.0.0,<5.0.0
# torch>=1.9.0,<3.0.0